refactor: add index export interfaces
