+++ src/components/User/PanelUsers.tsx
@@ -1,12 +1,12 @@
 import React, { useCallback, useEffect, useState } from 'react';
-import { IUser } from '@/utils/interfaces/interfaces';
+import { IUser } from '@/utils/interfaces';
 import styles from './User.module.scss';
 import { socketService } from '@/services/socketService';
 import { User } from './User';
 import { SocketEvent } from '@/utils/enums';
 import { UserModel } from '@/models/UserModel';
 
-export const PanelUsers = () => {
+export const PanelUsers = (): JSX.Element => {
   const [users, setUsers] = useState<IUser[]>([]);
 
   const subscribeJoin = useCallback((user: IUser) => {
